\name{menodata}
\docType{data}
\alias{menodata}
\title{Interval Censored Data for Age at Menopause}
\description{Interval censored data for age at menopause. Originally collected analyzed
as right censored data in MacMahon and Worcestor 1966. Krailo and Pike
1983 recommended treating as current status data to remove recall bias
evedent by terminal clustering around digits 5 and 10. Original data
also recorded whether menopause was natural or medically induced. 
This is left out in this dataset}
\format{A 2423 by 2 matrix in standard interval censored format}
\source{MacMahon and Worcestor 1966}
\references{Krailo and Pike 1983}